Pull out the new pass-style QuantumRuntime to new module path.
